+++ faddr/faddr_db.py
@@ -3,9 +3,12 @@
 import argparse
 import sys
 
+from rich.console import Console
+
 from faddr import logger
-from faddr.exceptions import FaddrSettingsFileFormatError
-from faddr.rancid import RancidDir
+from faddr.exceptions import FaddrParserUnknownProfile, FaddrSettingsFileFormatError
+from faddr.parser import Parser
+from faddr.rancid import RancidDir, RancidGroup
 from faddr.settings import load_settings
 
 
@@ -31,6 +34,10 @@ def parse_cmd_args():
 
 def main():
     """Parsing devices' config files and writing data to database."""
+
+    # Setup rich console for pretty printing
+    console = Console()
+
     cmd_args = parse_cmd_args()
     logger.debug(f"Arguments from CMD: {cmd_args}")
 
@@ -40,3 +47,23 @@ def main():
         logger.info(f"Failed to load settings: {err}")
         sys.exit(1)
     logger.debug(f"Generated settings: {settings.dict()}")
+
+    for rancid_dir in settings.rancid.dirs:
+        if rancid_dir.kind == "dir":
+            rancid = RancidDir(rancid_dir.path)
+        elif rancid_dir.kind in ("group", "repo"):
+            rancid = RancidGroup(rancid_dir.path)
+        else:
+            continue
+
+        for config in rancid.configs:
+            try:
+                parser = Parser(
+                    config["path"],
+                    rancid_dir.mapping.get(config["content_type"]),
+                    settings.templates_dir,
+                )
+                data = parser.parse()
+                console.print(data)
+            except FaddrParserUnknownProfile:
+                logger.debug(f"Unsupported config: {config}")